+++ releases.md
@@ -1,6 +1,6 @@
 # Releases
 
-## Unreleased
+## v0.55.6
 
   - Move Falcon middleware trace providers to `traces/provider/falcon/middleware`.
   - Add `traces` to the test dependencies for Falcon middleware trace provider tests.